Move noop_metrics back int metrics_lib to prevent import loops
