+++ mypy.ini
@@ -34,8 +34,5 @@ disallow_untyped_defs = True
 [mypy-paasta_tools.metrics.metrics_lib]
 disallow_untyped_defs = True
 
-[mypy-paasta_tools.metrics.noop_metrics]
-disallow_untyped_defs = True
-
 [mypy-paasta_tools.secret_tools]
 disallow_untyped_defs = True
