+++ paasta_tools/metrics/metrics_lib.py
@@ -1,4 +1,5 @@
 import logging
+import time
 from abc import ABC
 from abc import abstractmethod
 from typing import Any
@@ -24,15 +25,15 @@ _metrics_interfaces: Dict[str, Type['BaseMetrics']] = {}
 
 class TimerProtocol(Protocol):
     def start(self) -> None:
-        ...
+        raise NotImplementedError()
 
     def stop(self) -> None:
-        ...
+        raise NotImplementedError()
 
 
 class GaugeProtocol(Protocol):
     def set(self, value: Union[int, float]) -> None:
-        ...
+        raise NotImplementedError()
 
 
 class BaseMetrics(ABC):
@@ -41,11 +42,11 @@ class BaseMetrics(ABC):
 
     @abstractmethod
     def create_timer(self, name: str, **kwargs: Any) -> TimerProtocol:
-        ...
+        raise NotImplementedError()
 
     @abstractmethod
     def create_gauge(self, name: str, **kwargs: Any) -> GaugeProtocol:
-        ...
+        raise NotImplementedError()
 
 
 def get_metrics_interface(base_name: str) -> BaseMetrics:
@@ -72,3 +73,34 @@ class MeteoriteMetrics(BaseMetrics):
 
     def create_gauge(self, name: str, **kwargs: Any) -> GaugeProtocol:
         return yelp_meteorite.create_gauge(self.base_name + '.' + name, kwargs)
+
+
+class Timer(TimerProtocol):
+    def __init__(self, name: str) -> None:
+        self.name = name
+
+    def start(self) -> None:
+        log.debug("timer {} start at {}".format(self.name, time.time()))
+
+    def stop(self) -> None:
+        log.debug("timer {} stop at {}".format(self.name, time.time()))
+
+
+class Gauge(GaugeProtocol):
+    def __init__(self, name: str) -> None:
+        self.name = name
+
+    def set(self, value: Union[int, float]) -> None:
+        log.debug("gauge {} set to {}".format(self.name, value))
+
+
+@register_metrics_interface(None)
+class NoMetrics(BaseMetrics):
+    def __init__(self, base_name: str) -> None:
+        self.base_name = base_name
+
+    def create_timer(self, name: str, **kwargs: Any) -> Timer:
+        return Timer(self.base_name + '.' + name)
+
+    def create_gauge(self, name: str, **kwargs: Any) -> Gauge:
+        return Gauge(self.base_name + '.' + name)
