+++ tox.ini
@@ -141,7 +141,6 @@ mypy_paths =
     paasta_tools/secret_tools.py
     paasta_tools/utils.py
     paasta_tools/metrics/metrics_lib.py
-    paasta_tools/metrics/noop_metrics.py
     tests/deployd/test_common.py
     tests/deployd/test_watchers.py
     tests/frameworks
